feat(session-archive): reimplement navigation to add session screen
I've decided to redesign the behaviour of the session archive - now
when users click a list item, instead of seeing a context menu with
options for editing or deleting, they will be navigated to a "session
data" screen, where they can view the full data of a particular sleep
session, edit that data, or delete the session.

This commit represents the first in a series of commits that will
reimplement existing session archive behaviour with this new design.
Right now many tests have been commented out, I felt this was acceptable
because these tests once more represent future behaviour to be implemented.