feat(archive): restrict session creation/updates
This prevents the user from being able to create or edit sessions to
be in particular states. Users were already unable to set the start
time after the end time, or the end before the start, but now they are
also unable to:

- cause a session to overlap with another existing session (unless
    the user is editing that existing session)

- cause a session to happen in the future (this is to prevent
    real-time generated sessions (from the sleep tracking screen)
    creating overlaps)

Also:
The error message for start>end has been changed to a dialog, to be
consistent with the 2 new errors above which both use dialogs.