fix(tracker): fix session timer not resetting
This is a fix for a bug where the session timer would reset properly
after stopping a session, but then on app restart it would be as if
the session had never stopped.

The reason turned out to be missing a commit() call in CurrentSessionPrefs,
which caused the session prefs data to not actually get cleared!