feat(archive): implement interruption delete functionality
This is actually a pretty large commit, as to prepare for the
InterruptionDetailsFragment I redesigned SessionDetailsFragment so
they could share a common base (DetailsFragment).

I also removed the listeners from DetailsFragment, instead going with a
shared DetailsResult ViewModel system - this is because I think actually
the listeners were causing a bug, retaining a reference to the
parent fragment and not letting the framework deal with that
fragment properly.