fix(*): fix prepopulated database
When I originally extracted the db schema from the room json files,
I had included the setup commands for the room master table (inserting
the id & hash value). This was a mistake - this was "prepopulating"
the db with db version information and was causing migration problems.

Easy solution: I just removed the setup commands from the schema that
extract-schema.py generates and rebuilt prepopulated.db