fix(*): fix detail screen delete dialog crashes
Fragments deriving from DetailsFragment were crashing when mixing the delete dialog
with rotating the device. This was due to the delete dialog confirmation callback
trying to access DetailsFragment's activity too early after rotation. The solution was
to have the callback instead change the view model state, then observe that state change
once the DetailsFragment was ready.