fix(*): fix session times error dialogs
This relates to the "start < end" and "future time" session time
error dialogs.

Before, when you went to change a session time, if you confirmed an
invalid time in the picker AFTER rotating, the app would crash because
the error dialog was referring to the old fragment.

I created the "LiveDataEvent" system to fix this - now the error
dialogs are events triggered through the view model. (This system
should eventually be applied to all UI events in the app)