fix(interruptions): fix up/back button behaviour
This fixes a problem where pressing up or back on the interruptions details screen
still appeared to update the interruption in the sleep session details screen.

The problem was the Interruption reference used by the InterruptionDetailsFragment
being the same reference as the Interruption in the SleepSession's interruptions list.
I.e. a duplicate reference problem (or as I now want to call it: "entangled variables" :D)