fix(interruptions): fix interruptions deletion bug
This fixes a bug where, when editing a sleep session, you could add an interruption then
delete it, but when you confirm the sleep session that deleted interruption would still
be added.

The bug was due to the interruption being in both the 'added' and 'deleted' lists of
Interruptions.Updates, but receiving a new id when actually added to the db.

The solution was to remove the interruption from the 'added' list instead of ever
putting it in the 'deleted' list.