fix(stats): fix durations chart y labels
Fix for a problem where a very large sleep session (e.g. several years) would cause
the stats screen to have serious performance problems or crash.

This was due to the durations chart creating a y label for each hour of that session, so
I put a cap on the number of hours where the chart behaves that way.