fix(details): fix crash in the details when backgrounding the app
This fixes a bug where the app would crash when being backgrounded while
the Sleep Details screen was open. The problem was certain fields of
SleepSessionWrapper not being serializable (Interruptions, Mood).

The solution implemented here is to just have those classes implement
Serializable. This is just a quick patchup - it's not ideal and should
be returned to in the future. A better solution would be to leave these
core models pure and have parcelable wrappers for them at the
infrastructure level.