fix(tracker): fix mood selector crash
This fixes a crash that would happen if the app was sent to the
background while the mood selector dialog was open.

The problem was actually part of a systemic issue with the current
design of many parts of the app involving un-serializable data and
sending the app to the background.